Optimize output for list len()==1
Collapse mvfields with a length of 1 to a single value and set the value for
any empty list explicitly to None.  (No point is seeing "[]" in an output field
or an unnecessary wrapper list: `["value"]`.  If the user really *wants* such
behavior, just add a `| to_string(@)` on the end of the expression.